requirements.txt
selenium
beautifulsoup4
requests
python-dotenv
psycopg2-binary